daemon/requirements.txt
python-uinput
python-evdev